requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
idna==3.3
python-dotenv==0.19.2
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
